+++ ED/src/br/com/edu/ordenacao/Main.java
@@ -3,13 +3,8 @@ package br.com.edu.ordenacao;
 public class Main {
 
     public static void main(String[] args) {
-        int vetor[] = new int[]{3, 0, 1, 8, 7, 2, 5, 4, 9, 6};
+        int vetor[] = new int[]{3, 0, 1, 8, 7, 2, 5, 4, 9, 6,10,23,15,11,22,18,12,19,14};
 
-        QuickSort.ordena(vetor, 0, vetor.length - 1);
-
-        for (int i : vetor) {
-            System.out.print(i + " ");
-        }
-
+        SortingAlgorithms.quickSort(vetor, 0, vetor.length - 1);
     }
 }
